feat: Add "Item Category Ext" table extension
This commit adds a new table extension called "Item Category Ext" to the codebase. The table extension extends the "Item Category" table and includes a new field for "Item Category Type". This enhancement improves the functionality and usability of the application by providing more detailed information about the type of item category.