feat: Add "Game Name" field to "Item Card Extension"
This commit adds a new field called "Game Name" to the "Item Card Extension" page extension. The field is added to the existing "field" section and has the "ApplicationArea" set to "All". This enhancement improves the functionality and usability of the application by providing more detailed information about the game associated with an item.