feat: Add "Item Category Type" field to "Item Category Card" page extension
This commit adds a new field called "Item Category Type" to the "Item Category Card" page extension. The field is added after the "Parent Category" field and is set to be visible in all application areas. This addition enhances the functionality and usability of the application by allowing users to specify the type of item category.