feat: Add fields for "Task Nos" and "Project Mgt Nos" in "ExtPurchasesPayablesSetup" table
This commit adds two new fields, "Task Nos" and "Project Mgt Nos", to the "ExtPurchasesPayablesSetup" table. The "Task Nos" field is of type code[20] and is related to the "No. Series" table. The "Project Mgt Nos" field is also of type code[20] and is related to the "No. Series" table. Additionally, the caption of the "Project Mgt Nos" field is updated to 'Contract Nos'. These changes enhance the functionality and usability of the application by providing a structured way to store and manage task and project management numbers in the "ExtPurchasesPayablesSetup" table.